#430: Splitted up common.py
