+++ exasol_integration_test_docker_environment/cli/commands/spawn_test_environment.py
@@ -25,7 +25,7 @@ from exasol_integration_test_docker_environment.lib import api
 from exasol_integration_test_docker_environment.lib.api.api_errors import (
     ArgumentConstraintError,
 )
-from exasol_integration_test_docker_environment.lib.api.common import add_options
+from exasol_integration_test_docker_environment.lib.utils.cli_function_decorators import add_options
 
 
 @cli.command()
