+++ exasol_integration_test_docker_environment/lib/api/health.py
@@ -2,7 +2,7 @@ from inspect import cleandoc
 
 from exasol_integration_test_docker_environment.doctor import health_checkup
 from exasol_integration_test_docker_environment.lib.api.api_errors import HealthProblem
-from exasol_integration_test_docker_environment.lib.api.common import cli_function
+from exasol_integration_test_docker_environment.lib.utils.api_function_decorators import cli_function
 
 
 @cli_function
