+++ exasol_integration_test_docker_environment/lib/config/log_config.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 import luigi
