+++ exasol_integration_test_docker_environment/lib/logging/luigi_log_config.py
