+++ exasol_integration_test_docker_environment/lib/utils/api_function_decorators.py
@@ -1,0 +1,11 @@
+
+def cli_function(func):
+    """Decorator: Register a function as having a cli equivalent"""
+    func.__cli_function__ = True
+    return func
+
+
+def no_cli_function(func):
+    """Decorator: Register a function as noy having a cli equivalent"""
+    func.__cli_function__ = False
+    return func
