+++ exasol_integration_test_docker_environment/lib/utils/cli_function_decorators.py
@@ -1,0 +1,8 @@
+
+def add_options(options):
+    def _add_options(func):
+        for option in reversed(options):
+            func = option(func)
+        return func
+
+    return _add_options
