+++ exasol_integration_test_docker_environment/test/test_docker_access_method.py
@@ -6,7 +6,6 @@ from typing import (
 import luigi
 import pytest
 
-from exasol_integration_test_docker_environment.lib.api.common import generate_root_task
 from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
     DependencyLoggerBaseTask,
 )
