+++ exasol_integration_test_docker_environment/test/test_populate_data.py
@@ -7,7 +7,7 @@ from sys import stderr
 
 import luigi
 
-from exasol_integration_test_docker_environment.lib.api.common import generate_root_task
+from exasol_integration_test_docker_environment.lib.api.run_task import generate_root_task
 from exasol_integration_test_docker_environment.lib.data.test_container_content_description import (
     TestContainerRuntimeMapping,
 )
