+++ exasol_integration_test_docker_environment/test/test_test_container_reuse.py
@@ -5,14 +5,12 @@ from pathlib import Path
 
 import luigi
 
-from exasol_integration_test_docker_environment.lib.api.common import (
-    generate_root_task,
-    set_build_config,
-    set_docker_repository_config,
-)
 from exasol_integration_test_docker_environment.lib.base.docker_base_task import (
     DockerBaseTask,
 )
+from exasol_integration_test_docker_environment.lib.api.run_task import generate_root_task
+from exasol_integration_test_docker_environment.lib.config.build_config import set_build_config
+from exasol_integration_test_docker_environment.lib.config.docker_config import set_docker_repository_config
 from exasol_integration_test_docker_environment.lib.data.container_info import (
     ContainerInfo,
 )
